Change input types from int to float

+++ calculator.py
@@ -1,5 +1,5 @@
-x = int(input("Enter a number: "))
-y = int(input("Enter another number: "))
+x = float(input("Enter a number: "))
+y = float(input("Enter another number: "))
 print("Addition: " , x+y)
 print("Subtraction: " , x-y)
 print("Multiplication: " , x*y)
